傾きとコンテンツの透明度を連動する。
