+++ lib/home_ui_model.freezed.dart
@@ -26,12 +26,12 @@ $HomeUIModelCopyWith<HomeUIModel> get copyWith => _$HomeUIModelCopyWithImpl<Home
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is HomeUIModel&&const DeepCollectionEquality().equals(other.members, members)&&(identical(other.isInAnimation, isInAnimation) || other.isInAnimation == isInAnimation)&&(identical(other.animationDuration, animationDuration) || other.animationDuration == animationDuration)&&(identical(other.width, width) || other.width == width)&&(identical(other.height, height) || other.height == height)&&(identical(other.startDragX, startDragX) || other.startDragX == startDragX)&&(identical(other.startDragY, startDragY) || other.startDragY == startDragY)&&(identical(other.cardAppearance, cardAppearance) || other.cardAppearance == cardAppearance)&&(identical(other.animationBeginCardAppearance, animationBeginCardAppearance) || other.animationBeginCardAppearance == animationBeginCardAppearance));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is HomeUIModel&&const DeepCollectionEquality().equals(other.members, members)&&(identical(other.isInAnimation, isInAnimation) || other.isInAnimation == isInAnimation)&&(identical(other.animationDuration, animationDuration) || other.animationDuration == animationDuration)&&(identical(other.width, width) || other.width == width)&&(identical(other.height, height) || other.height == height)&&(identical(other.startDragX, startDragX) || other.startDragX == startDragX)&&(identical(other.startDragY, startDragY) || other.startDragY == startDragY)&&const DeepCollectionEquality().equals(other.cardAppearance, cardAppearance)&&const DeepCollectionEquality().equals(other.animationBeginCardAppearance, animationBeginCardAppearance));
 }
 
 
 @override
-int get hashCode => Object.hash(runtimeType,const DeepCollectionEquality().hash(members),isInAnimation,animationDuration,width,height,startDragX,startDragY,cardAppearance,animationBeginCardAppearance);
+int get hashCode => Object.hash(runtimeType,const DeepCollectionEquality().hash(members),isInAnimation,animationDuration,width,height,startDragX,startDragY,const DeepCollectionEquality().hash(cardAppearance),const DeepCollectionEquality().hash(animationBeginCardAppearance));
 
 @override
 String toString() {
@@ -50,7 +50,7 @@ $Res call({
 });
 
 
-$CardAppearanceCopyWith<$Res> get cardAppearance;$CardAppearanceCopyWith<$Res> get animationBeginCardAppearance;
+
 
 }
 /// @nodoc
@@ -63,7 +63,7 @@ class _$HomeUIModelCopyWithImpl<$Res>
 
 /// Create a copy of HomeUIModel
 /// with the given fields replaced by the non-null parameter values.
-@pragma('vm:prefer-inline') @override $Res call({Object? members = null,Object? isInAnimation = null,Object? animationDuration = null,Object? width = null,Object? height = null,Object? startDragX = null,Object? startDragY = null,Object? cardAppearance = null,Object? animationBeginCardAppearance = null,}) {
+@pragma('vm:prefer-inline') @override $Res call({Object? members = null,Object? isInAnimation = null,Object? animationDuration = null,Object? width = null,Object? height = null,Object? startDragX = null,Object? startDragY = null,Object? cardAppearance = freezed,Object? animationBeginCardAppearance = freezed,}) {
   return _then(_self.copyWith(
 members: null == members ? _self.members : members // ignore: cast_nullable_to_non_nullable
 as List<Member>,isInAnimation: null == isInAnimation ? _self.isInAnimation : isInAnimation // ignore: cast_nullable_to_non_nullable
@@ -72,30 +72,12 @@ as Duration,width: null == width ? _self.width : width // ignore: cast_nullable_
 as double,height: null == height ? _self.height : height // ignore: cast_nullable_to_non_nullable
 as double,startDragX: null == startDragX ? _self.startDragX : startDragX // ignore: cast_nullable_to_non_nullable
 as double,startDragY: null == startDragY ? _self.startDragY : startDragY // ignore: cast_nullable_to_non_nullable
-as double,cardAppearance: null == cardAppearance ? _self.cardAppearance : cardAppearance // ignore: cast_nullable_to_non_nullable
-as CardAppearance,animationBeginCardAppearance: null == animationBeginCardAppearance ? _self.animationBeginCardAppearance : animationBeginCardAppearance // ignore: cast_nullable_to_non_nullable
+as double,cardAppearance: freezed == cardAppearance ? _self.cardAppearance : cardAppearance // ignore: cast_nullable_to_non_nullable
+as CardAppearance,animationBeginCardAppearance: freezed == animationBeginCardAppearance ? _self.animationBeginCardAppearance : animationBeginCardAppearance // ignore: cast_nullable_to_non_nullable
 as CardAppearance,
   ));
 }
-/// Create a copy of HomeUIModel
-/// with the given fields replaced by the non-null parameter values.
-@override
-@pragma('vm:prefer-inline')
-$CardAppearanceCopyWith<$Res> get cardAppearance {
-  
-  return $CardAppearanceCopyWith<$Res>(_self.cardAppearance, (value) {
-    return _then(_self.copyWith(cardAppearance: value));
-  });
-}/// Create a copy of HomeUIModel
-/// with the given fields replaced by the non-null parameter values.
-@override
-@pragma('vm:prefer-inline')
-$CardAppearanceCopyWith<$Res> get animationBeginCardAppearance {
-  
-  return $CardAppearanceCopyWith<$Res>(_self.animationBeginCardAppearance, (value) {
-    return _then(_self.copyWith(animationBeginCardAppearance: value));
-  });
-}
+
 }
 
 
@@ -132,12 +114,12 @@ _$HomeUIModelCopyWith<_HomeUIModel> get copyWith => __$HomeUIModelCopyWithImpl<_
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is _HomeUIModel&&const DeepCollectionEquality().equals(other._members, _members)&&(identical(other.isInAnimation, isInAnimation) || other.isInAnimation == isInAnimation)&&(identical(other.animationDuration, animationDuration) || other.animationDuration == animationDuration)&&(identical(other.width, width) || other.width == width)&&(identical(other.height, height) || other.height == height)&&(identical(other.startDragX, startDragX) || other.startDragX == startDragX)&&(identical(other.startDragY, startDragY) || other.startDragY == startDragY)&&(identical(other.cardAppearance, cardAppearance) || other.cardAppearance == cardAppearance)&&(identical(other.animationBeginCardAppearance, animationBeginCardAppearance) || other.animationBeginCardAppearance == animationBeginCardAppearance));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _HomeUIModel&&const DeepCollectionEquality().equals(other._members, _members)&&(identical(other.isInAnimation, isInAnimation) || other.isInAnimation == isInAnimation)&&(identical(other.animationDuration, animationDuration) || other.animationDuration == animationDuration)&&(identical(other.width, width) || other.width == width)&&(identical(other.height, height) || other.height == height)&&(identical(other.startDragX, startDragX) || other.startDragX == startDragX)&&(identical(other.startDragY, startDragY) || other.startDragY == startDragY)&&const DeepCollectionEquality().equals(other.cardAppearance, cardAppearance)&&const DeepCollectionEquality().equals(other.animationBeginCardAppearance, animationBeginCardAppearance));
 }
 
 
 @override
-int get hashCode => Object.hash(runtimeType,const DeepCollectionEquality().hash(_members),isInAnimation,animationDuration,width,height,startDragX,startDragY,cardAppearance,animationBeginCardAppearance);
+int get hashCode => Object.hash(runtimeType,const DeepCollectionEquality().hash(_members),isInAnimation,animationDuration,width,height,startDragX,startDragY,const DeepCollectionEquality().hash(cardAppearance),const DeepCollectionEquality().hash(animationBeginCardAppearance));
 
 @override
 String toString() {
@@ -156,7 +138,7 @@ $Res call({
 });
 
 
-@override $CardAppearanceCopyWith<$Res> get cardAppearance;@override $CardAppearanceCopyWith<$Res> get animationBeginCardAppearance;
+
 
 }
 /// @nodoc
@@ -169,7 +151,7 @@ class __$HomeUIModelCopyWithImpl<$Res>
 
 /// Create a copy of HomeUIModel
 /// with the given fields replaced by the non-null parameter values.
-@override @pragma('vm:prefer-inline') $Res call({Object? members = null,Object? isInAnimation = null,Object? animationDuration = null,Object? width = null,Object? height = null,Object? startDragX = null,Object? startDragY = null,Object? cardAppearance = null,Object? animationBeginCardAppearance = null,}) {
+@override @pragma('vm:prefer-inline') $Res call({Object? members = null,Object? isInAnimation = null,Object? animationDuration = null,Object? width = null,Object? height = null,Object? startDragX = null,Object? startDragY = null,Object? cardAppearance = freezed,Object? animationBeginCardAppearance = freezed,}) {
   return _then(_HomeUIModel(
 members: null == members ? _self._members : members // ignore: cast_nullable_to_non_nullable
 as List<Member>,isInAnimation: null == isInAnimation ? _self.isInAnimation : isInAnimation // ignore: cast_nullable_to_non_nullable
@@ -178,160 +160,9 @@ as Duration,width: null == width ? _self.width : width // ignore: cast_nullable_
 as double,height: null == height ? _self.height : height // ignore: cast_nullable_to_non_nullable
 as double,startDragX: null == startDragX ? _self.startDragX : startDragX // ignore: cast_nullable_to_non_nullable
 as double,startDragY: null == startDragY ? _self.startDragY : startDragY // ignore: cast_nullable_to_non_nullable
-as double,cardAppearance: null == cardAppearance ? _self.cardAppearance : cardAppearance // ignore: cast_nullable_to_non_nullable
-as CardAppearance,animationBeginCardAppearance: null == animationBeginCardAppearance ? _self.animationBeginCardAppearance : animationBeginCardAppearance // ignore: cast_nullable_to_non_nullable
+as double,cardAppearance: freezed == cardAppearance ? _self.cardAppearance : cardAppearance // ignore: cast_nullable_to_non_nullable
+as CardAppearance,animationBeginCardAppearance: freezed == animationBeginCardAppearance ? _self.animationBeginCardAppearance : animationBeginCardAppearance // ignore: cast_nullable_to_non_nullable
 as CardAppearance,
-  ));
-}
-
-/// Create a copy of HomeUIModel
-/// with the given fields replaced by the non-null parameter values.
-@override
-@pragma('vm:prefer-inline')
-$CardAppearanceCopyWith<$Res> get cardAppearance {
-  
-  return $CardAppearanceCopyWith<$Res>(_self.cardAppearance, (value) {
-    return _then(_self.copyWith(cardAppearance: value));
-  });
-}/// Create a copy of HomeUIModel
-/// with the given fields replaced by the non-null parameter values.
-@override
-@pragma('vm:prefer-inline')
-$CardAppearanceCopyWith<$Res> get animationBeginCardAppearance {
-  
-  return $CardAppearanceCopyWith<$Res>(_self.animationBeginCardAppearance, (value) {
-    return _then(_self.copyWith(animationBeginCardAppearance: value));
-  });
-}
-}
-
-/// @nodoc
-mixin _$CardAppearance {
-
- double get offsetX; double get offsetY; double get angle;
-/// Create a copy of CardAppearance
-/// with the given fields replaced by the non-null parameter values.
-@JsonKey(includeFromJson: false, includeToJson: false)
-@pragma('vm:prefer-inline')
-$CardAppearanceCopyWith<CardAppearance> get copyWith => _$CardAppearanceCopyWithImpl<CardAppearance>(this as CardAppearance, _$identity);
-
-
-
-@override
-bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is CardAppearance&&(identical(other.offsetX, offsetX) || other.offsetX == offsetX)&&(identical(other.offsetY, offsetY) || other.offsetY == offsetY)&&(identical(other.angle, angle) || other.angle == angle));
-}
-
-
-@override
-int get hashCode => Object.hash(runtimeType,offsetX,offsetY,angle);
-
-@override
-String toString() {
-  return 'CardAppearance(offsetX: $offsetX, offsetY: $offsetY, angle: $angle)';
-}
-
-
-}
-
-/// @nodoc
-abstract mixin class $CardAppearanceCopyWith<$Res>  {
-  factory $CardAppearanceCopyWith(CardAppearance value, $Res Function(CardAppearance) _then) = _$CardAppearanceCopyWithImpl;
-@useResult
-$Res call({
- double offsetX, double offsetY, double angle
-});
-
-
-
-
-}
-/// @nodoc
-class _$CardAppearanceCopyWithImpl<$Res>
-    implements $CardAppearanceCopyWith<$Res> {
-  _$CardAppearanceCopyWithImpl(this._self, this._then);
-
-  final CardAppearance _self;
-  final $Res Function(CardAppearance) _then;
-
-/// Create a copy of CardAppearance
-/// with the given fields replaced by the non-null parameter values.
-@pragma('vm:prefer-inline') @override $Res call({Object? offsetX = null,Object? offsetY = null,Object? angle = null,}) {
-  return _then(_self.copyWith(
-offsetX: null == offsetX ? _self.offsetX : offsetX // ignore: cast_nullable_to_non_nullable
-as double,offsetY: null == offsetY ? _self.offsetY : offsetY // ignore: cast_nullable_to_non_nullable
-as double,angle: null == angle ? _self.angle : angle // ignore: cast_nullable_to_non_nullable
-as double,
-  ));
-}
-
-}
-
-
-/// @nodoc
-
-
-class _CardAppearance implements CardAppearance {
-  const _CardAppearance({required this.offsetX, required this.offsetY, required this.angle});
-  
-
-@override final  double offsetX;
-@override final  double offsetY;
-@override final  double angle;
-
-/// Create a copy of CardAppearance
-/// with the given fields replaced by the non-null parameter values.
-@override @JsonKey(includeFromJson: false, includeToJson: false)
-@pragma('vm:prefer-inline')
-_$CardAppearanceCopyWith<_CardAppearance> get copyWith => __$CardAppearanceCopyWithImpl<_CardAppearance>(this, _$identity);
-
-
-
-@override
-bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is _CardAppearance&&(identical(other.offsetX, offsetX) || other.offsetX == offsetX)&&(identical(other.offsetY, offsetY) || other.offsetY == offsetY)&&(identical(other.angle, angle) || other.angle == angle));
-}
-
-
-@override
-int get hashCode => Object.hash(runtimeType,offsetX,offsetY,angle);
-
-@override
-String toString() {
-  return 'CardAppearance(offsetX: $offsetX, offsetY: $offsetY, angle: $angle)';
-}
-
-
-}
-
-/// @nodoc
-abstract mixin class _$CardAppearanceCopyWith<$Res> implements $CardAppearanceCopyWith<$Res> {
-  factory _$CardAppearanceCopyWith(_CardAppearance value, $Res Function(_CardAppearance) _then) = __$CardAppearanceCopyWithImpl;
-@override @useResult
-$Res call({
- double offsetX, double offsetY, double angle
-});
-
-
-
-
-}
-/// @nodoc
-class __$CardAppearanceCopyWithImpl<$Res>
-    implements _$CardAppearanceCopyWith<$Res> {
-  __$CardAppearanceCopyWithImpl(this._self, this._then);
-
-  final _CardAppearance _self;
-  final $Res Function(_CardAppearance) _then;
-
-/// Create a copy of CardAppearance
-/// with the given fields replaced by the non-null parameter values.
-@override @pragma('vm:prefer-inline') $Res call({Object? offsetX = null,Object? offsetY = null,Object? angle = null,}) {
-  return _then(_CardAppearance(
-offsetX: null == offsetX ? _self.offsetX : offsetX // ignore: cast_nullable_to_non_nullable
-as double,offsetY: null == offsetY ? _self.offsetY : offsetY // ignore: cast_nullable_to_non_nullable
-as double,angle: null == angle ? _self.angle : angle // ignore: cast_nullable_to_non_nullable
-as double,
   ));
 }
