+++ test/home_ui_model_test.dart
@@ -1,4 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:skip_like_flutter/card_appearance.dart';
 import 'package:skip_like_flutter/home_ui_model.dart';
 import 'package:skip_like_flutter/model/member.dart';
